(feature) split  files and app.py
