+++ _UI/_web_interface/app.py
@@ -1,0 +1,28 @@
+# KrakenSDR Signal Processor
+#
+# Copyright (C) 2018-2021  Carl Laufer, Tamás Pető
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+#
+# - coding: utf-8 -*-
+
+from app_layout import make_layout
+from maindash import app
+
+if __name__ == "__main__":
+    app.layout = make_layout()
+    # Debug mode does not work when the data interface is set to shared-memory
+    # "shmem"!
+    app.run_server(debug=False, host="0.0.0.0", port=8080)
