+++ kill.sh
@@ -8,6 +8,6 @@ else
     KILL_SIGNAL=64
 fi
 
-sudo kill -${KILL_SIGNAL} $(ps ax | grep "[p]ython3 _UI/_web_interface/kraken_web_interface.py" | awk '{print $1}') 2> /dev/null
+sudo kill -${KILL_SIGNAL} $(ps ax | grep "[p]ython3 _UI/_web_interface/app.py" | awk '{print $1}') 2> /dev/null
 sudo kill -${KILL_SIGNAL} $(ps ax | grep "[p]hp" | awk '{print $1}') 2> /dev/null
 sudo pkill -f node
